Fix Call Api
se arregla la llamada a la api con el pair

+++ app/src/main/java/conversormoneda/api_service/ApiRecord.java
@@ -3,5 +3,10 @@ package conversormoneda.api_service;
 /**
  * ApiRecord
  */
-public record ApiRecord(String result, String base_code, String target_code, int conversion_rate, int conversion_result) {
+public record ApiRecord(String result,
+                        String base_code,
+                        String target_code,
+                        double conversion_rate,
+                        double conversion_result) {
+
 }
